Update README.md initial startup instructions
The initial startup instructions say that wp-env start can be used after changing into a directory that contains a WordPress plugin.  In my testing, although wp-env start will succeed without a .wp-env.json file, wp-env will not successfully detect that it is in a plugin directory (or that it is in the root of a WordPress installation for that matter) without a .wp-env.json file.  

Rather, it reports  "Warning: could not find a .wp-env.json configuration file and could not determine if '/Users/nachbar/work/2024/progs/wordpress/test03/wordpress/wp-content/plugins/contact-form-7' is a WordPress installation, a plugin, or a theme."

Therefore, so first-time-users will be able to successfully follow the instructions, they must include information about creating a .wp-env.json file.

Also, the fact that a different environment is created depending on the directory from which wp-env is run is a significant block for the first-time user, so including that information, and how to see the docker mappings, is also added.